Disable GeoIP enrichment
When sending an event to PostHog, the server will "enrich" that event with GeoIP location data. However, for client libraries like this that are meant to be used on the server, that doesn't make much sense since it'll always be the location data of the server hosting this client.

This sets the `$geoip_disable` property to `true` which makes reading information about captured events from this library much easier.

If someone wants GeoIP enrichment, they can open an issue and I'll consider it.

Also bumped the version